pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Add your description here"
authors = ["Erick Marinho <erickgaves@gmail.com>"]
readme = "README.md"
# Defina suas dependências de desenvolvimento aqui, se houver
# packages = [{include = "app"}] # Descomente se 'app' for seu pacote principal

[tool.poetry.dependencies]
python = ">=3.12, <4.0"
fastapi = {version = ">=0.115.12", extras = ["standard"]}
pydantic-ai = ">=0.1.10"
python-dotenv = ">=1.1.0"
uvicorn = {version = ">=0.34.2,<0.35.0", extras = ["standard"]}
langgraph = ">=0.4.3"
langgraph-checkpoint-postgres = ">=2.0.21"
psycopg = {version = ">=3.2.8", extras = ["binary"]}
psycopg-pool = ">=3.2.6"
langchain-core = "^0.3.59"
langchain-openai = "^0.3.16"

[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
authors = [
    {name = "Erick Marinho", email = "erickgaves@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
# dependencies = [] # Comente ou remova esta linha da seção [project]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Se você tiver dependências de desenvolvimento (ex: pytest, ruff)
